:building_construction: crop 로직 위치 수정, cleaningCode
