+++ app/src/main/res/layout/view_section_selected_label.xml
@@ -51,6 +51,5 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintBottom_toBottomOf="parent"
-        android:layout_marginHorizontal="20dp"
         android:background="@color/colorGrey4" />
 </merge>
