+++ app/src/main/res/values/strings.xml
@@ -66,10 +66,14 @@
     <string name="btn_apply">"등록"</string>
     <string name="tv_more">"…더보기"</string>
 
+    <string name="tag_label">태그 추가</string>
     <string name="tag_edit_hint">#태그 입력하기</string>
     <string name="tag_max_count">#최대 10개까지 등록</string>
     <string name="tag_regex">#%1$s</string>
+    <string name="upload_label">씨향지 만들기</string>
     <string name="upload_only_show_to_me">나만 보기</string>
+    <string name="upload_perfume_section_title">향수 선택</string>
+    <string name="upload_tag_section_title">태그 추가</string>
     <string name="upload_title">씨향지 만들기</string>
     <string name="editor_failed_save_edited_image">이미지 저장에 실패했습니다..</string>
 </resources>
